feat(pgsql): support postgresql in php scripts that check connection and craft installation at startup
see #42
